Upgrade to Spring Boot 1.3.2.RELEASE
* Use latest stable release of Spring Boot
* Upgrade to gradle wrapper 2.9
* Rename source code to package io.springoneplatform
* Add license to source code
* Plug in @SpringBootApplication annotation
