--- README.md
@@ -1,8 +1,0 @@
-platform-java
-======================
-
-```
-gradle build
-java -jar build/libs/site-0.1.0.jar or gradle run or cf push
-```
-http://localhost:8080/
